requirements.txt
gunicorn==20.1.0
Flask==2.0.3
pandas==1.4.1
Pillow==9.1.0
regex==2021.10.8
requests==2.26.0
Werkzeug==2.0.3
